Modularized gdown codebase
